requirements.txt
Django==1.11.2
djangorestframework==3.6.3
pytz==2017.2
pycrypto==2.6.1
